backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
bson==0.5.10
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.63.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==3.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
uuid==1.30
Werkzeug==3.1.3
